[Debug] Add file emitter for tywaves
Added support for emitting a new file format suitable for the tywaves-surfer waveform viewer. This file format has been created since HGLDD wasn't really suitable for the typed viewer.
This should be considered as an alpha version for prototyping and its actual utilization next to HGLDD needs to be discussed. In addition, since good part of the code is shared with the hgldd emitter, it's likely that the source lang information will by included within HGLDD: having a bigger unique file format should be preferred than having a simpler second file format.

This file can be emitted using the `--emit-tywaves-hgldd` command line option in firtool.